moved thumbnail preview functions to a separate previewFunctions.py
