+++ functions/__init__.py
@@ -1,2 +1,3 @@
 from .mainFunctions import *
 from .jsonFunctions import *
+from .previewFunctions import *
